src/pyproject.toml
[project]
name = "53-54-booking-cafe-seats-team1"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Golovushkin",email = "frenkodetto@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.119)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "fastapi-mail (==1.4.2)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.2"
pre-commit = "^4.3.0"
alembic = "^1.17.0"
